+++ CHANGELOG.md
@@ -23,6 +23,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+* Methods in the `analyzer` module now return backend-agnostic tensors.
 * Add `IndependentMultivariateDistribution` for per-channel distribution
   definition.
 * Add automatic knot selection public API to core library.